moved index.js login component into it's own component. created routes to work with app
